main/requirements.txt
torch
numpy
fastapi
pydantic
Pillow
pandas
streamlit
scipy
matplotlib
seaborn
uvicorn
tqdm
scikit-learn